Added frontend and backend for Dhuttle tracking Feature

+++ backend/src/main/java/com/campus360/repository/ShuttleRouteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShuttleRouteRepository extends JpaRepository<ShuttleRoute, Long> {
+public interface ShuttleRouteRepository extends JpaRepository<ShuttleRoute, Integer> {
 }
